test: add tests for other controller actions

+++ backend/src/main/java/com/fgc/combo/companion/service/impl/PlaylistServiceImpl.java
@@ -1,14 +1,5 @@
 package com.fgc.combo.companion.service.impl;
 
-import java.net.URLDecoder;
-import java.nio.charset.StandardCharsets;
-import java.util.List;
-
-import org.springframework.beans.BeanUtils;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
-
 import com.fgc.combo.companion.dto.AddCombosToPlaylistDTO;
 import com.fgc.combo.companion.dto.CreatePlaylistDTO;
 import com.fgc.combo.companion.dto.PaginationResponse;
@@ -23,8 +14,14 @@ import com.fgc.combo.companion.repository.PlaylistRepository;
 import com.fgc.combo.companion.service.PlaylistComboService;
 import com.fgc.combo.companion.service.PlaylistService;
 import com.fgc.combo.companion.service.UserService;
-
 import jakarta.transaction.Transactional;
+import java.net.URLDecoder;
+import java.nio.charset.StandardCharsets;
+import java.util.List;
+import org.springframework.beans.BeanUtils;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
 
 @Service
 public class PlaylistServiceImpl implements PlaylistService {
@@ -34,50 +31,70 @@ public class PlaylistServiceImpl implements PlaylistService {
   private final PlaylistComboService playlistComboService;
 
   public PlaylistServiceImpl(
-      PlaylistRepository playlistRepository,
-      UserService userService,
-      PlaylistComboService playlistComboService) {
+    PlaylistRepository playlistRepository,
+    UserService userService,
+    PlaylistComboService playlistComboService
+  ) {
     this.playlistRepository = playlistRepository;
     this.userService = userService;
     this.playlistComboService = playlistComboService;
   }
 
-  private Playlist getByIdAndOwner(Long playlistId, User user) {
-    return this.playlistRepository.findByIdAndOwner(playlistId, user)
-        .orElseThrow(() -> new ResourceNotFoundException("Playlist not found!"));
+  private Playlist getByIdWithOwnerValidation(Long playlistId) {
+    User currentUser = userService.me();
+    Playlist playlist = getPlaylistById(playlistId);
+
+    if (playlist.getOwner().getId() != currentUser.getId()) {
+      throw new OperationNotAllowedException(
+        "This playlist belongs to another user!"
+      );
+    }
+
+    return playlist;
   }
 
   private Playlist getPlaylistById(Long id) {
     return this.playlistRepository.findById(id)
-        .orElseThrow(() -> new ResourceNotFoundException("Playlist not found!"));
+      .orElseThrow(() -> new ResourceNotFoundException("Playlist not found!"));
   }
 
   @Override
   public Playlist addCombosToPlaylist(
-      Long playlistId,
-      AddCombosToPlaylistDTO combos) {
+    Long playlistId,
+    AddCombosToPlaylistDTO combos
+  ) {
     Playlist playlist = getPlaylistById(playlistId);
     this.playlistComboService.addAllCombosToPlaylist(
         playlist,
-        combos.getCombos());
+        combos.getCombos()
+      );
     return playlist;
   }
 
   @Override
   public Playlist getPlaylistWithCombos(Long playlistId) {
     return this.playlistRepository.findById(playlistId)
-        .orElseThrow(() -> new ResourceNotFoundException("Playlist not found!"));
+      .orElseThrow(() -> new ResourceNotFoundException("Playlist not found!"));
   }
 
   @Transactional
   @Override
   public boolean deleteCombosFromPlaylist(
-      Long playlistId,
-      List<Long> playlistComboIds) {
-    Playlist playlist = getByIdAndOwner(playlistId, this.userService.me());
+    Long playlistId,
+    List<Long> playlistComboIds
+  ) {
+    User currentUser = userService.me();
+    Playlist playlist = getPlaylistById(playlistId);
+
+    if (playlist.getOwner().getId() != currentUser.getId()) {
+      throw new OperationNotAllowedException(
+        "This playlist belongs to another user!"
+      );
+    }
     this.playlistComboService.removeCombosFromPlaylist(
         playlist,
-        playlistComboIds);
+        playlistComboIds
+      );
     return true;
   }
 
@@ -94,22 +111,15 @@ public class PlaylistServiceImpl implements PlaylistService {
 
     this.playlistComboService.addAllCombosToPlaylist(
         createdPlaylist,
-        playlistDTO.getCombos());
+        playlistDTO.getCombos()
+      );
 
     return createdPlaylist;
   }
 
   @Override
   public Playlist update(Long id, UpdatePlaylistDTO playlistDTO) {
-    User currentUser = userService.me();
-
-    Playlist playlist = playlistRepository
-        .findById(id)
-        .orElseThrow(() -> new ResourceNotFoundException("Playlist not found!"));
-
-    if (playlist.getOwner().getId() != currentUser.getId())
-      throw new OperationNotAllowedException(
-          "This playlist belongs to another user!");
+    Playlist playlist = getByIdAndCurrentUser(id);
 
     BeanUtils.copyProperties(playlistDTO, playlist);
     return this.savePlaylist(playlist);
@@ -117,20 +127,20 @@ public class PlaylistServiceImpl implements PlaylistService {
 
   @Override
   public Playlist getByIdAndCurrentUser(Long id) {
-    return getByIdAndOwner(id, this.userService.me());
+    return getByIdWithOwnerValidation(id);
   }
 
   @Override
   public PaginationResponse<Playlist> getByCurrentUser(Pageable pageable) {
-    Page<Playlist> currentUserPlaylists = this.playlistRepository.findByOwner(this.userService.me(), pageable);
+    Page<Playlist> currentUserPlaylists =
+      this.playlistRepository.findByOwner(this.userService.me(), pageable);
     return PaginationResponseMapper.create(currentUserPlaylists);
   }
 
   @Override
   @Transactional
   public boolean delete(Long playlistId) {
-    Playlist playlist = getByIdAndOwner(playlistId, this.userService.me());
-
+    Playlist playlist = getByIdWithOwnerValidation(playlistId);
     this.playlistComboService.deleteByPlaylist(playlist);
     this.playlistRepository.delete(playlist);
 
@@ -144,11 +154,17 @@ public class PlaylistServiceImpl implements PlaylistService {
 
   @Override
   public PaginationResponse<Playlist> getAllByPlaylistNameOrTagName(
-      PlaylistComboSearchDTO playlistComboResponseDTO,
-      Pageable pageable) {
+    PlaylistComboSearchDTO playlistComboResponseDTO,
+    Pageable pageable
+  ) {
     return PaginationResponseMapper.create(
-        this.playlistRepository.findAllByComboNameAndTagName(
-            URLDecoder.decode(playlistComboResponseDTO.getName(), StandardCharsets.UTF_8),
-            pageable));
+      this.playlistRepository.findAllByComboNameAndTagName(
+          URLDecoder.decode(
+            playlistComboResponseDTO.getName(),
+            StandardCharsets.UTF_8
+          ),
+          pageable
+        )
+    );
   }
 }
